New projection in same slot fix rebuild
Now when a new projection is in the same slot of existing
projection, before a rebuild, the value of last sent checkpoint
is fixed to the value of the entire slot.

Still we can have problem for new projection in new slot
because in nitro mode they will not flush.