Projection Engine: New Slot are not rebuilded
If you start projection engine with a new projection in a new
slot, it will start from 0, but if you starts in rebuild mode
it never get StopRebuilding called.

With this fix, new projection in new slot, are not run in rebuild
mode even if rebuild is set to true.