Fixed bug of snapshot and serialization of Repo.
First we disabled the dictionary that contains loaded snapshot for
aggregate, because the snapshot is passed as is to the aggregate
and can be modfiied. If a concurrency exception occours the snapshot
that was potentially modified by the aggregate was incorrectly returned
to the new aggreate.
Moreover there are conccurent exception potentially on accessing
collection for aggregate state, so corruption occurred.

Second fix is the addition of an Id to the repository to be double
sure that we remove the lock on an entity only by the repository
that really acquired it, thus minimizing the concurrent exception.

Release of the lock on entity was put in a finally block, because
for concurrency it was not correctly released, resulting in lock
being held until timeout (5 seconds approx).