Persister and on-the-fly Aggregate Query Model
Sometimes there is the need to do a on-the-fly projection
of an aggregate given its events.
In some situation it is not safe to access a ReadModel
because it can be Stale and there is the need to have a quick
way to create a projection of a given aggregate. Thanks to the
Persister<> class, we can create persister that are able to
read all the events of an Aggregate and create an on the fly projection.

We also added the concept of ISnapshotPerister, a component that
is capable to persist snapshots on some medium.