Refactoring of BusBootsrapper + JarvisStartable
Refactoring of BusBootstrapper to be used as Startable facility
as all the other components. This is allowed because of
JarvisStartable facility that allows for explicit startup
and priority in startable.

The initialization of rebus is also separated in two
distinct phases. The original BusBootstrapper only manage
registration of IBus and configuration.

You need also to register the BusStarter that is the component
that is actually used to start the bus. With this trick
you can register and start bus during different moment
of the Istartable pattern.