Offline capabilities (squash of 35 commits)
The two major changes are various modification
and new classes to support offline functionalities
and also the ability to have non domain events
in commits.

------------------COMMENTS-------------------

Change version signature.

Because Jarvis already used an aggregate with state
with a property Signature that conflicted with the
base property of the aggregate State.

Added IEngine settings for hosts.

Changed tracking for continuous rebuild.

In standard tracking during projections all commits with
zero events are not dispatched, and the count remains equal
to zero.

We need to always update the slot if we are in continuous rebulid.

Added support for non DomainEvents in stream

We can have some commits with payload that
are not domain events, this happens if we use
NES as simple Stream without an aggregate.

Some modification are needed to the framework
because some dispatchers always assume that
EventStream contains only DomainEvents.

Fixed logic of CommitEnhancer

The Commit enhancer incorrectly cast the id to
EventStoreIdentity to an unused variables.

This prevent storing in the stream commit that
have messages not inheriting from DomainEvents
that have StreamId of a class that does not inherits
from EventStoreIdentity