Refactoring for IConcurrentCheckpointTracker
Extracted a different component called ISlotStatusManager
to manage slots using a new Attribute that should be present
in all projection.

This attribute is used to specify Projection: commonName
slotName and signature without the need to instantiate an
instance of the projection.

This will allow Projection Engine to check for slot status
(new, need to be rebuild, missing) directly before any
initialization, to appropriately configure the rebuild engine.