Ported framework to NSTORE
+semver: breaking
+semver: major

This is a combination of more than 90
original commits that were done to
gradually migrate the Jarvis.Framework
to NStore engine.

In this specific commit we squashed also
some of the fixes we did on async projection
to fix bugs that were discovered while usage
with NStore but not really related to Nstore.

We decided that it would be more useful to
have a single commit than a sequence of micro
commits that are really more difficult to follow
since migration to NStore really changed the
very basic on how the framework works.