Fix for offline - online collection wrapper idempotency.
In previous fix, we expect the event to contain an header with
the array of all offline events. When offline/online sync is performed,
this array is passed in command context and serialized to string.

The fix consist in deserialize the array if it was previously serialized
to correctly check for idempotency.