Optimization for concurrent checkpoint tracker.
Previous version is simple, whenever the projection engine
dispatched a chunk, it will immediately update all checkpoint
record in mongo database with the new id of dispatched chunk.

This creates lots of unnecessary work, because for most slots
when you dispatch a slot, the projection did not processed the
event, leading to lots of unnecessary checkpoint updates.

To avoid this, ConcurrentCheckpointTracker was not updated
to update record in database only if the slot really processed
the event or a specific timeout is passed.