Fix interrupted rebuild bug #11561
This was a really insidious bug because it presents almost never
and generates only side effect.

Start a rebuild and then interrupt the rebuild.
You know that you should restart rebuidl, but if you start projection
engine in normal mode, it will simple re-project every slot that was
interrupted during rebuild.

This causes
1. All process mangaer to rerun
2. not using nitro for rebuild
3. general slowness of the system.